exFat: Fix the size of the cluster_number type
ClusterChain::next's FAT read was reading 64 bits of data from the FAT and then slicing into the wrong 32 bits, because the cluster_number type was incorrectly defined as a std::uint64_t.

Also perform a case-insensitive match of file name extensions in the sio file system device.

+++ src/fs/exfat/AddressTypes.h
@@ -5,7 +5,7 @@
 
 namespace fs { namespace exfat {
 
-using cluster_number = std::uint64_t;
+using cluster_number = std::uint32_t;
 
 using logical_cluster_number = cluster_number;
 using virtual_cluster_number = cluster_number;
